Rename async/sync to asynchronous/synchronous
This is needed since 'async' became a keyword in py37
