+++ adbts/tcp/__init__.py
@@ -16,6 +16,6 @@ tcp_timeout = functools.partial(timeouts.timeout,  # pylint: disable=invalid-nam
                                 seconds=True)
 
 
-from . import async, sync  # noqa pylint: disable=wrong-import-position
+from . import asynchronous, synchronous  # noqa pylint: disable=wrong-import-position
 
-__all__ = ['sync', 'async']
+__all__ = ['asynchronous', 'synchronous']
