+++ adbts/tcp/asynchronous.py
@@ -1,11 +1,10 @@
 """
-    adbts.tcp.sync
-    ~~~~~~~~~~~~~~
+    adbts.tcp.asynchronous
+    ~~~~~~~~~~~~~~~~~~~~~~
 
-    Contains functionality for synchronous Transmission Control Protocol (TCP) transport.
+    Contains functionality for asynchronous Transmission Control Protocol (TCP) transport using `asyncio`.
 """
-import contextlib
-import socket
+import asyncio
 
 from . import tcp_timeout
 from .. import exceptions, hints, timeouts, transport
@@ -13,35 +12,21 @@ from .. import exceptions, hints, timeouts, transport
 __all__ = ['Transport']
 
 
-@contextlib.contextmanager
-def socket_timeout_scope(sock: hints.Socket, timeout: hints.Timeout = timeouts.UNDEFINED):
-    """
-    Patches the socket timeout for the scope of the context manager.
-
-    :param sock: Socket to set the timeout on
-    :type sock: :class:`~socket.socket`
-    :param timeout: Socket timeout in milliseconds
-    :type timeout: :class:`~int`
-    """
-    current_timeout = sock.gettimeout()
-    sock.settimeout(timeout)
-    try:
-        yield
-    finally:
-        sock.settimeout(current_timeout)
-
+# Disable incorrect warning on asyncio.wait_for, https://github.com/PyCQA/pylint/issues/996.
+# pylint: disable=not-an-iterable
 
 class Transport(transport.Transport):
     """
-    Defines synchronous (blocking) TCP transport.
-
-    .. note:: This transport is not thread-safe.
+    Defines asynchronous (non-blocking) TCP transport using `asyncio`.
     """
 
-    def __init__(self, host: hints.Str, port: hints.Int, sock: hints.Socket) -> None:
+    def __init__(self, host: hints.Str, port: hints.Int, reader: hints.StreamReader,
+                 writer: hints.StreamWriter, loop: hints.EventLoop) -> None:
         self._host = host
         self._port = port
-        self._socket = sock
+        self._reader = reader
+        self._writer = writer
+        self._loop = loop
 
     def __repr__(self):
         return '<{}(address={!r}, state={!r})>'.format(self.__class__.__name__, str(self),
@@ -58,12 +43,13 @@ class Transport(transport.Transport):
         :return: Closed state of the transport
         :rtype: :class:`~bool`
         """
-        return self._socket is None or self._socket._closed  # pylint: disable=protected-access
+        return self._reader is None or self._writer is None
 
+    @asyncio.coroutine
     @transport.ensure_opened
     @transport.ensure_num_bytes
     @exceptions.reraise(OSError)
-    @exceptions.reraise_timeout_errors(socket.timeout)
+    @exceptions.reraise_timeout_errors(asyncio.TimeoutError)
     def read(self, num_bytes: hints.Int,
              timeout: hints.Timeout = timeouts.UNDEFINED) -> transport.TransportReadResult:
         """
@@ -78,13 +64,16 @@ class Transport(transport.Transport):
         :raises :class:`~adbts.exceptions.TransportError`: When underlying transport encounters an error
         :raises :class:`~adbts.exceptions.TimeoutError`: When timeout is exceeded
         """
-        with socket_timeout_scope(self._socket, tcp_timeout(timeout)):
-            return self._socket.recv(num_bytes)
+        data = yield from asyncio.wait_for(self._reader.read(num_bytes),
+                                           timeout=tcp_timeout(timeout),
+                                           loop=self._loop)
+        return data
 
+    @asyncio.coroutine
     @transport.ensure_opened
     @transport.ensure_data
     @exceptions.reraise(OSError)
-    @exceptions.reraise_timeout_errors(socket.timeout)
+    @exceptions.reraise_timeout_errors(asyncio.TimeoutError)
     def write(self, data: hints.Buffer,
               timeout: hints.Timeout = timeouts.UNDEFINED) -> transport.TransportWriteResult:
         """
@@ -95,13 +84,12 @@ class Transport(transport.Transport):
         :param timeout: Maximum number of milliseconds to write before raising an exception.
         :type timeout: :class:`~int`, :class:`~NoneType`, or :class:`~object`
         :return Nothing
-        :rtype: :class:`~NoneType`
-        :raises :class:`~adbts.exceptions.TransportError`: When underlying transport encounters an error
+        :return: :class:`~NoneType`
+        :raises :class:`~adbts.exceptions.TransportError`: When underlying transport encounters an error.
         :raises :class:`~adbts.exceptions.TimeoutError`: When timeout is exceeded
         """
-        with socket_timeout_scope(self._socket, tcp_timeout(timeout)):
-            self._socket.sendall(data)
-            return None
+        self._writer.write(data)
+        yield from asyncio.wait_for(self._writer.drain(), timeout=tcp_timeout(timeout), loop=self._loop)
 
     @transport.ensure_opened
     @exceptions.reraise(OSError)
@@ -113,27 +101,31 @@ class Transport(transport.Transport):
         :rtype: `None`
         :raises :class:`~adbts.exceptions.TransportError`: When underlying transport encounters an error
         """
-        self._socket.close()
-        self._socket = None
+        self._writer.close()
+        self._writer = None
+        self._reader = None
 
 
+@asyncio.coroutine
 @exceptions.reraise(OSError)
-@exceptions.reraise_timeout_errors(socket.timeout)
 def open(host: hints.Str, port: hints.Int,  # pylint: disable=redefined-builtin
-         timeout: hints.Timeout = timeouts.UNDEFINED) -> transport.TransportOpenResult:
+         timeout: hints.Timeout = timeouts.UNDEFINED,
+         loop: hints.EventLoop = None) -> transport.TransportOpenResult:
     """
-    Open a new :class:`~adbts.tcp.sync.Transport` transport to the given host/port.
+    Open a new :class:`~adbts.tcp.async.Transport` transport to the given host/port.
 
     :param host: Remote host
     :type host: :class:`~str`
     :param port: Remote port
     :type port: :class:`~int`
-    :param timeout: Maximum number of milliseconds on blocking socket operations before raising an exception
+    :param timeout: Maximum number of milliseconds to write before raising an exception.
     :type timeout: :class:`~int`, :class:`~NoneType`, or :class:`~object`
-    :return: Synchronous TCP transport
-    :rtype: :class:`~adbts.tcp.sync.Transport`
+    :param loop: Asyncio Event Loop
+    :type loop: :class:`~asyncio.events.AbstractEventLoop`
+    :return: Asynchronous TCP transport
+    :rtype: :class:`~adbts.tcp.async.Transport`
     :raises :class:`~adbts.exceptions.TransportError`: When underlying transport encounters an error
-    :raises :class:`~adbts.exceptions.TimeoutError`: When timeout is exceeded
     """
-    sock = socket.create_connection((host, port), tcp_timeout(timeout))
-    return Transport(host, port, sock)
+    reader, writer = yield from asyncio.wait_for(asyncio.open_connection(host, port, loop=loop),
+                                                 timeout=tcp_timeout(timeout), loop=loop)
+    return Transport(host, port, reader, writer, loop)
