+++ adbts/usb/asynchronous.py
@@ -1,6 +1,6 @@
 """
-    adbts.usb.async
-    ~~~~~~~~~~~~~~~
+    adbts.usb.asynchronous
+    ~~~~~~~~~~~~~~~~~~~~~~
 
     Contains functionality for asynchronous Universal Serial Bus (USB) transport.
 """
