--- tests/tcp/test_tcp_async.py
@@ -1,14 +1,0 @@
-"""
-    test_tcp_async
-    ~~~~~~~~~~~~~~
-
-    Tests for the :mod:`~adbts.tpc.async` module.
-"""
-import pytest
-
-from adbts.tcp import async
-
-
-@pytest.mark.xfail(reason='Not implemented')
-def test_stub():
-    assert False
