+++ tests/tcp/test_tcp_asynchronous.py
@@ -1,0 +1,14 @@
+"""
+    test_tcp_asynchronous
+    ~~~~~~~~~~~~~~~~~~~~~
+
+    Tests for the :mod:`~adbts.tpc.asynchronous` module.
+"""
+import pytest
+
+from adbts.tcp import asynchronous
+
+
+@pytest.mark.xfail(reason='Not implemented')
+def test_stub():
+    assert False
