--- tests/tcp/test_tcp_sync.py
@@ -1,14 +1,0 @@
-"""
-    test_tcp_sync
-    ~~~~~~~~~~~~~
-
-    Tests for the :mod:`~adbts.tpc.sync` module.
-"""
-import pytest
-
-from adbts.tcp import sync
-
-
-@pytest.mark.xfail(reason='Not implemented')
-def test_stub():
-    assert False
