+++ tests/tcp/test_tcp_synchronous.py
@@ -1,0 +1,14 @@
+"""
+    test_tcp_synchronous
+    ~~~~~~~~~~~~~~~~~~~~
+
+    Tests for the :mod:`~adbts.tpc.synchronous` module.
+"""
+import pytest
+
+from adbts.tcp import synchronous
+
+
+@pytest.mark.xfail(reason='Not implemented')
+def test_stub():
+    assert False
